Remove "verify media stream URL" feature
The library will no longer return the content-length or the
content-type (or its derivate: file extension) values along
with the media properties. This was done by sending an HTTP
HEAD request to the server and parsing the values in question
from the reply metadata.

This saves bandwidth, and allows extending the media scripts to
return all available media stream URLs (e.g. with >1 formats)
to the library.

Applications that need this, may do so with the media stream URL
they choose to use using the quvi_verify_* function set or use a
different solution altogether. This patch adds the quvi_verify_*
function set.