Remove 'query formats' feature
Querying the available formats for a media URL will no longer be
supported by the library the way it was implemented in 0.4 . The new
design of the media script support that allows returning all available
media stream URLs for different available formats with just one call,
obsoletes the "traditional query formats" implementation found in 0.4 .

This patch removes the planned support for the quvi_query_formats_*
function set. It also also obsoletes the support for the "query_formats"
function in the media scripts.