Replace qargs.stream_url with qargs.streams
This commit changes fundamentally how the media scripts return the
parsed properties back to the library.

Instead of returning only one media stream URL, all of them are
returned in a dictionary, immediately accessible to the applications
using the library.

This means also that the applications will no longer have to query the
available formats first, set the requested format and then have the
media script parse the media properties again, is a thing of the past.

There are additional media (stream) properties available, e.g. video
height, width, encoding, etc. when applicable (iow: when they are
returned by the server and may be parsed in a media script).

It should be noted, however, that with libquvi 1.0, the only mandatory
media property is the media stream URL. There must be at least one of
those returned in the qargs.streams.

These changes complete the removal of the "verification of media
stream URL" that took place after parsing a media stream URL.