Add quvi_media_stream_select
This patch adds the quvi_media_stream_select function which may be used
to select the stream by its format ID from the available streams. It
supports patterns, comma-separated values and adds support for the
'croak' keyword.

This feature was initially created to allow users specify a list of
formats instead of only one, with a bailout method ('croak') to generate
an error if nothing matched.

Originally this was intended to be added to the quvi/util.lua:choose_format
function but was later replaced with a separate C API function to complement
the qargs.stream support in libquvi 1.0 .