Comment out l_quvi_object_opts_curl
This function currently sets only the CURLOPT_USERAGENT. The scripts may
use the quvi.http.header to set the UA which is more flexible: the
custom headers will be cleared automatically, thus the previously set
CURLOPT_USERAGENT will not be lost, e.g. when a program is reusing the
libquvi curl session handle and sets the CURLOPT_USERAGENT to another
value.

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>